+++ tests/assets/ignore2.ts
@@ -1,0 +1,9 @@
+export interface MyApi {
+  foo: string;
+  bar: boolean;
+}
+
+/**
+ * @removeprop bar
+ */
+export interface MyApi {}