+++ tests/ignores.test.ts
@@ -23,3 +23,26 @@ test('should not drop ignored properties if noIgnore set to true', () => {
   }
 }`);
 });
+
+test('should remove prop in interface merging', () => {
+  const result = runTestFor('ignore2.ts');
+  expect(result).toBe(`declare module "test" {
+  export interface MyApi {
+    foo: string;
+  }
+}`);
+});
+
+test('should remove cause in interface merging', () => {
+  const result = runTestFor('ignore3.ts');
+  expect(result).toBe(`declare module "test" {
+  /**
+   * This comment should still exist.
+   */
+  export interface MyApi extends MyApi2 {}
+
+  export interface MyApi2 {
+    bar: boolean;
+  }
+}`);
+});